Rename to clarify that it processes the tutorials
